Mover input container a un widget
